Move FitFlightSeatView to separate file
